[Refactor] User-Refrigerator 엔티티 간 연관관계 재설정
- 애플리케이션에 가입한 '회원'을 지칭하는 엔티티의 이름을 User로 수정 (기존 엔티티명: Member)
    - 패키지 명을 member에서 user로 변경
- User와 Refrigerator의 관계를 다대다 관계로 수정
    - User와 Refrigerator 사이에 Member라는 중간 엔티티를 두고 각 엔티티와 일대다 관계로 연결
- Member의 권한을 결정하는 MemberRole enum 클래스 생성 (STAFF, MEMBER)
