+++ src/main/kotlin/team/b2/bingojango/domain/member/model/MemberRole.kt
@@ -1,5 +1,5 @@
 package team.b2.bingojango.domain.member.model
 
-enum class UserRole {
-    ADMIN, USER
+enum class MemberRole {
+    STAFF, MEMBER
 }
