+++ src/main/kotlin/team/b2/bingojango/domain/member/repository/MemberRepository.kt
@@ -2,10 +2,8 @@ package team.b2.bingojango.domain.member.repository
 
 import org.springframework.data.jpa.repository.JpaRepository
 import org.springframework.stereotype.Repository
-import team.b2.bingojango.domain.member.model.User
+import team.b2.bingojango.domain.member.model.Member
 
 @Repository
-interface UserRepository : JpaRepository<User, Long> {
-
-    fun findByEmail(email:String) : User?
+interface MemberRepository : JpaRepository<Member, Long> {
 }
