+++ src/main/kotlin/team/b2/bingojango/domain/user/controller/UserController.kt
@@ -1,4 +1,4 @@
-package team.b2.bingojango.domain.member.controller
+package team.b2.bingojango.domain.user.controller
 
 import org.springframework.http.HttpStatus
 import org.springframework.http.ResponseEntity
@@ -7,9 +7,9 @@ import org.springframework.web.bind.annotation.PostMapping
 import org.springframework.web.bind.annotation.RequestBody
 import org.springframework.web.bind.annotation.RequestMapping
 import org.springframework.web.bind.annotation.RestController
-import team.b2.bingojango.domain.member.dto.LoginRequest
-import team.b2.bingojango.domain.member.dto.LoginResponse
-import team.b2.bingojango.domain.member.service.UserService
+import team.b2.bingojango.domain.user.dto.LoginRequest
+import team.b2.bingojango.domain.user.dto.LoginResponse
+import team.b2.bingojango.domain.user.service.UserService
 
 
 @RestController
