+++ src/main/kotlin/team/b2/bingojango/domain/user/dto/LoginRequest.kt
@@ -1,4 +1,4 @@
-package team.b2.bingojango.domain.member.dto
+package team.b2.bingojango.domain.user.dto
 
 data class LoginRequest (
     val email: String,
