+++ src/main/kotlin/team/b2/bingojango/domain/user/dto/LoginResponse.kt
@@ -1,4 +1,4 @@
-package team.b2.bingojango.domain.member.dto
+package team.b2.bingojango.domain.user.dto
 
 data class LoginResponse (
     val accessToken: String
