+++ src/main/kotlin/team/b2/bingojango/domain/user/model/UserRole.kt
@@ -1,0 +1,5 @@
+package team.b2.bingojango.domain.user.model
+
+enum class UserRole {
+    ADMIN, USER
+}
