+++ src/main/kotlin/team/b2/bingojango/domain/user/model/UserStatus.kt
@@ -1,4 +1,4 @@
-package team.b2.bingojango.domain.member.model
+package team.b2.bingojango.domain.user.model
 
 enum class UserStatus {
     NORMAL, BANNED, WITHDRAWN
